+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Draft
+
+### Added
+
+Added styles for consent banner buttons for mobiles [#7](https://github.com/bigcommerce/consent-manager/pull/7)
+
 ## 4.1.0(Dec 11, 2019)
 
 ### Added